+++ backend/src/main.rs
@@ -1,13 +1,13 @@
-use actix_web::{web, App, HttpResponse, HttpServer, Responder, HttpRequest};
-use actix_web_actors::ws;
-use actix::{Actor, StreamHandler, Addr, Message, AsyncContext};
+use actix::{Actor, Addr, AsyncContext, Message, StreamHandler};
 use actix_cors::Cors;
-use rusqlite::{params, Connection};
-use serde::{Deserialize, Serialize};
-use std::sync::{Mutex, Arc};
-use std::collections::HashMap;
+use actix_web::{App, HttpRequest, HttpResponse, HttpServer, Responder, web};
+use actix_web_actors::ws;
 use evalexpr::*;
 use regex;
+use rusqlite::{Connection, params};
+use serde::{Deserialize, Serialize};
+use std::collections::HashMap;
+use std::sync::{Arc, Mutex};
 use uuid::Uuid;
 
 pub struct AppState {
@@ -17,6 +17,7 @@ pub struct AppState {
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 struct Cell {
+    sheet: Option<String>,
     row: i32,
     col: i32,
     value: String,
@@ -29,6 +30,7 @@ struct Cell {
 #[derive(Message, Serialize, Deserialize, Clone)]
 #[rtype(result = "()")]
 pub struct CellUpdate {
+    pub sheet: String,
     pub row: i32,
     pub col: i32,
     pub value: String,
@@ -62,11 +64,13 @@ impl Actor for WebSocketSession {
     fn started(&mut self, ctx: &mut Self::Context) {
         let mut sessions = self.sessions.lock().unwrap();
         sessions.insert(self.id.clone(), ctx.address());
-        
+
         // Notify other users that a new user joined
-        let msg = UserJoined { user_id: self.id.clone() };
+        let msg = UserJoined {
+            user_id: self.id.clone(),
+        };
         let msg_str = serde_json::to_string(&msg).unwrap();
-        
+
         for (session_id, addr) in sessions.iter() {
             if session_id != &self.id {
                 addr.do_send(WebSocketMessage(msg_str.clone()));
@@ -77,11 +81,13 @@ impl Actor for WebSocketSession {
     fn stopped(&mut self, _ctx: &mut Self::Context) {
         let mut sessions = self.sessions.lock().unwrap();
         sessions.remove(&self.id);
-        
+
         // Notify other users that user left
-        let msg = UserLeft { user_id: self.id.clone() };
+        let msg = UserLeft {
+            user_id: self.id.clone(),
+        };
         let msg_str = serde_json::to_string(&msg).unwrap();
-        
+
         for (_, addr) in sessions.iter() {
             addr.do_send(WebSocketMessage(msg_str.clone()));
         }
@@ -107,8 +113,8 @@ impl StreamHandler<Result<ws::Message, ws::ProtocolError>> for WebSocketSession 
                         }
                     }
                 }
-            },
-            Ok(ws::Message::Binary(_)) => {},
+            }
+            Ok(ws::Message::Binary(_)) => {}
             _ => {}
         }
     }
@@ -122,7 +128,7 @@ impl actix::Handler<WebSocketMessage> for WebSocketSession {
     }
 }
 
-fn eval_formula(expr: &str, db_conn: &Connection) -> Result<String, String> {
+fn eval_formula(expr: &str, sheet: &str, db_conn: &Connection) -> Result<String, String> {
     let expr = expr.trim_start_matches('=');
     let mut ctx = HashMapContext::new();
     ctx.set_function(
@@ -157,20 +163,24 @@ fn eval_formula(expr: &str, db_conn: &Connection) -> Result<String, String> {
     for cap in re.captures_iter(expr) {
         let col_str = cap.get(1).map_or("", |m| m.as_str());
         let row_str = cap.get(2).map_or("", |m| m.as_str());
-        
-        let col = col_str.chars().fold(0, |acc, c| acc * 26 + (c as i32 - 'A' as i32 + 1)) - 1;
+
+        let col = col_str
+            .chars()
+            .fold(0, |acc, c| acc * 26 + (c as i32 - 'A' as i32 + 1))
+            - 1;
         let row = row_str.parse::<i32>().unwrap() - 1;
 
-        let mut stmt = db_conn.prepare("SELECT value FROM cells WHERE row = ?1 AND col = ?2").unwrap();
-        let value: Result<String, _> = stmt.query_row(params![row, col], |r| r.get(0));
-        
+        let mut stmt = db_conn
+            .prepare("SELECT value FROM cells WHERE sheet = ?1 AND row = ?2 AND col = ?3")
+            .unwrap();
+        let value: Result<String, _> = stmt.query_row(params![sheet, row, col], |r| r.get(0));
+
         if let Ok(val) = value {
             if let Ok(num) = val.parse::<f64>() {
                 final_expr = final_expr.replace(&cap[0], &num.to_string());
             }
         }
     }
-
 
     eval_with_context(&final_expr, &ctx)
         .map(|v| v.to_string())
@@ -181,18 +191,23 @@ async fn health() -> impl Responder {
     HttpResponse::Ok().body("ok")
 }
 
-async fn list_cells(data: web::Data<AppState>) -> impl Responder {
+async fn list_cells(
+    data: web::Data<AppState>,
+    query: web::Query<HashMap<String, String>>,
+) -> impl Responder {
+    let sheet = query.get("sheet").map(|s| s.as_str()).unwrap_or("default");
     let conn = data.db.lock().unwrap();
-    let mut stmt = conn.prepare("SELECT row, col, value, font_weight, font_style, background_color FROM cells").unwrap();
+    let mut stmt = conn.prepare("SELECT sheet, row, col, value, font_weight, font_style, background_color FROM cells WHERE sheet = ?1").unwrap();
     let rows = stmt
-        .query_map([], |r| {
+        .query_map(params![sheet], |r| {
             Ok(Cell {
-                row: r.get(0)?,
-                col: r.get(1)?,
-                value: r.get(2)?,
-                font_weight: r.get(3)?,
-                font_style: r.get(4)?,
-                background_color: r.get(5)?,
+                sheet: Some(r.get(0)?),
+                row: r.get(1)?,
+                col: r.get(2)?,
+                value: r.get(3)?,
+                font_weight: r.get(4)?,
+                font_style: r.get(5)?,
+                background_color: r.get(6)?,
             })
         })
         .unwrap();
@@ -206,70 +221,84 @@ async fn list_cells(data: web::Data<AppState>) -> impl Responder {
 async fn set_cell(data: web::Data<AppState>, item: web::Json<Cell>) -> impl Responder {
     let conn = data.db.lock().unwrap();
     let mut cell_to_save = item.clone();
+    let sheet = cell_to_save
+        .sheet
+        .clone()
+        .unwrap_or_else(|| "default".to_string());
+    cell_to_save.sheet = Some(sheet.clone());
 
     if cell_to_save.value.starts_with('=') {
-        if let Ok(res) = eval_formula(&cell_to_save.value, &conn) {
+        if let Ok(res) = eval_formula(&cell_to_save.value, &sheet, &conn) {
             cell_to_save.value = res;
         }
     }
     conn.execute(
-        "INSERT INTO cells (row, col, value, font_weight, font_style, background_color) 
-         VALUES (?1, ?2, ?3, ?4, ?5, ?6)
-         ON CONFLICT(row, col) DO UPDATE SET 
+        "INSERT INTO cells (sheet, row, col, value, font_weight, font_style, background_color)
+         VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7)
+         ON CONFLICT(sheet, row, col) DO UPDATE SET
             value=excluded.value,
             font_weight=excluded.font_weight,
             font_style=excluded.font_style,
             background_color=excluded.background_color",
         params![
-            cell_to_save.row, 
-            cell_to_save.col, 
+            cell_to_save.sheet,
+            cell_to_save.row,
+            cell_to_save.col,
             cell_to_save.value,
             cell_to_save.font_weight,
             cell_to_save.font_style,
             cell_to_save.background_color,
         ],
-    ).unwrap();
-    
+    )
+    .unwrap();
+
     // Broadcast the update to all connected WebSocket sessions
     broadcast_cell_update(&data.sessions, &cell_to_save, "system".to_string());
-    
+
     HttpResponse::Ok().body("saved")
 }
 
 async fn set_cells_bulk(data: web::Data<AppState>, items: web::Json<Vec<Cell>>) -> impl Responder {
     let conn = data.db.lock().unwrap();
-    
+
     // Start transaction for better performance
     let tx = conn.unchecked_transaction().unwrap();
-    
+
     for item in items.iter() {
         let mut cell_to_save = item.clone();
+        let sheet = cell_to_save
+            .sheet
+            .clone()
+            .unwrap_or_else(|| "default".to_string());
+        cell_to_save.sheet = Some(sheet.clone());
 
         if cell_to_save.value.starts_with('=') {
-            if let Ok(res) = eval_formula(&cell_to_save.value, &conn) {
+            if let Ok(res) = eval_formula(&cell_to_save.value, &sheet, &conn) {
                 cell_to_save.value = res;
             }
         }
-        
+
         tx.execute(
-            "INSERT INTO cells (row, col, value, font_weight, font_style, background_color) 
-             VALUES (?1, ?2, ?3, ?4, ?5, ?6)
-             ON CONFLICT(row, col) DO UPDATE SET 
+            "INSERT INTO cells (sheet, row, col, value, font_weight, font_style, background_color)
+             VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7)
+             ON CONFLICT(sheet, row, col) DO UPDATE SET
                 value=excluded.value,
                 font_weight=excluded.font_weight,
                 font_style=excluded.font_style,
                 background_color=excluded.background_color",
             params![
-                cell_to_save.row, 
-                cell_to_save.col, 
+                sheet,
+                cell_to_save.row,
+                cell_to_save.col,
                 cell_to_save.value,
                 cell_to_save.font_weight,
                 cell_to_save.font_style,
                 cell_to_save.background_color,
             ],
-        ).unwrap();
+        )
+        .unwrap();
     }
-    
+
     tx.commit().unwrap();
     HttpResponse::Ok().body("saved")
 }
@@ -277,11 +306,13 @@ async fn set_cells_bulk(data: web::Data<AppState>, items: web::Json<Vec<Cell>>) 
 #[derive(Serialize, Deserialize)]
 struct EvalRequest {
     expr: String,
+    sheet: Option<String>,
 }
 
 async fn evaluate(query: web::Json<EvalRequest>, data: web::Data<AppState>) -> impl Responder {
     let conn = data.db.lock().unwrap();
-    match eval_formula(&query.expr, &conn) {
+    let sheet = query.sheet.as_deref().unwrap_or("default");
+    match eval_formula(&query.expr, sheet, &conn) {
         Ok(result) => HttpResponse::Ok().body(result),
         Err(e) => HttpResponse::BadRequest().body(e),
     }
@@ -294,23 +325,28 @@ struct ClearRequest {
 
 #[derive(Serialize, Deserialize)]
 struct CellPosition {
+    sheet: Option<String>,
     row: i32,
     col: i32,
 }
 
-async fn clear_cells_bulk(data: web::Data<AppState>, request: web::Json<ClearRequest>) -> impl Responder {
+async fn clear_cells_bulk(
+    data: web::Data<AppState>,
+    request: web::Json<ClearRequest>,
+) -> impl Responder {
     let conn = data.db.lock().unwrap();
-    
+
     // Start transaction for better performance
     let tx = conn.unchecked_transaction().unwrap();
-    
+
     for pos in request.cells.iter() {
         tx.execute(
-            "DELETE FROM cells WHERE row = ?1 AND col = ?2",
-            params![pos.row, pos.col],
-        ).unwrap();
+            "DELETE FROM cells WHERE sheet = ?1 AND row = ?2 AND col = ?3",
+            params![pos.sheet.as_deref().unwrap_or("default"), pos.row, pos.col],
+        )
+        .unwrap();
     }
-    
+
     tx.commit().unwrap();
     HttpResponse::Ok().body("cleared")
 }
@@ -318,25 +354,29 @@ async fn clear_cells_bulk(data: web::Data<AppState>, request: web::Json<ClearReq
 pub fn init_db(conn: &Connection) {
     conn.execute(
         "CREATE TABLE IF NOT EXISTS cells (
+            sheet TEXT NOT NULL DEFAULT 'default',
             row INTEGER,
             col INTEGER,
             value TEXT,
             font_weight TEXT,
             font_style TEXT,
             background_color TEXT,
-            PRIMARY KEY (row, col)
+            PRIMARY KEY (sheet, row, col)
         )",
         [],
     )
     .unwrap();
 }
 
-
 // WebSocket endpoint
-async fn ws_index(req: HttpRequest, stream: web::Payload, data: web::Data<AppState>) -> Result<HttpResponse, actix_web::Error> {
+async fn ws_index(
+    req: HttpRequest,
+    stream: web::Payload,
+    data: web::Data<AppState>,
+) -> Result<HttpResponse, actix_web::Error> {
     let session_id = Uuid::new_v4().to_string();
     let sessions = data.sessions.clone();
-    
+
     ws::start(
         WebSocketSession {
             id: session_id,
@@ -348,8 +388,13 @@ async fn ws_index(req: HttpRequest, stream: web::Payload, data: web::Data<AppSta
 }
 
 // Helper function to broadcast cell updates
-fn broadcast_cell_update(sessions: &Arc<Mutex<HashMap<String, Addr<WebSocketSession>>>>, cell: &Cell, user_id: String) {
+fn broadcast_cell_update(
+    sessions: &Arc<Mutex<HashMap<String, Addr<WebSocketSession>>>>,
+    cell: &Cell,
+    user_id: String,
+) {
     let update = CellUpdate {
+        sheet: cell.sheet.clone().unwrap_or_else(|| "default".to_string()),
         row: cell.row,
         col: cell.col,
         value: cell.value.clone(),
@@ -358,7 +403,7 @@ fn broadcast_cell_update(sessions: &Arc<Mutex<HashMap<String, Addr<WebSocketSess
         background_color: cell.background_color.clone(),
         user_id,
     };
-    
+
     if let Ok(msg_str) = serde_json::to_string(&update) {
         let sessions_guard = sessions.lock().unwrap();
         for (_, addr) in sessions_guard.iter() {
@@ -372,7 +417,10 @@ async fn main() -> std::io::Result<()> {
     let conn = Connection::open("cells.db").unwrap();
     init_db(&conn);
 
-    let data = web::Data::new(AppState { db: Mutex::new(conn), sessions: Arc::new(Mutex::new(HashMap::new())) });
+    let data = web::Data::new(AppState {
+        db: Mutex::new(conn),
+        sessions: Arc::new(Mutex::new(HashMap::new())),
+    });
 
     HttpServer::new(move || {
         App::new()
@@ -383,7 +431,7 @@ async fn main() -> std::io::Result<()> {
                     .allowed_origin("http://192.168.10.161:3000")
                     .allow_any_method()
                     .allow_any_header()
-                    .supports_credentials()
+                    .supports_credentials(),
             )
             .app_data(data.clone())
             .route("/health", web::get().to(health))
@@ -409,7 +457,10 @@ mod tests {
     async fn health_works() {
         let conn = Connection::open_in_memory().unwrap();
         init_db(&conn);
-        let data = web::Data::new(AppState { db: Mutex::new(conn), sessions: Arc::new(Mutex::new(HashMap::new())) });
+        let data = web::Data::new(AppState {
+            db: Mutex::new(conn),
+            sessions: Arc::new(Mutex::new(HashMap::new())),
+        });
         let app = test::init_service(
             App::new()
                 .app_data(data.clone())
@@ -428,7 +479,10 @@ mod tests {
     async fn create_and_list_cells() {
         let conn = Connection::open_in_memory().unwrap();
         init_db(&conn);
-        let data = web::Data::new(AppState { db: Mutex::new(conn), sessions: Arc::new(Mutex::new(HashMap::new())) });
+        let data = web::Data::new(AppState {
+            db: Mutex::new(conn),
+            sessions: Arc::new(Mutex::new(HashMap::new())),
+        });
         let app = test::init_service(
             App::new()
                 .app_data(data.clone())
@@ -438,7 +492,15 @@ mod tests {
         )
         .await;
 
-        let new_cell = Cell { row: 1, col: 1, value: "42".into(), font_weight: None, font_style: None, background_color: None };
+        let new_cell = Cell {
+            sheet: Some("test".into()),
+            row: 1,
+            col: 1,
+            value: "42".into(),
+            font_weight: None,
+            font_style: None,
+            background_color: None,
+        };
         let req = test::TestRequest::post()
             .uri("/cells")
             .set_json(&new_cell)
@@ -446,7 +508,9 @@ mod tests {
         let resp = test::call_service(&app, req).await;
         assert!(resp.status().is_success());
 
-        let req = test::TestRequest::get().uri("/cells").to_request();
+        let req = test::TestRequest::get()
+            .uri("/cells?sheet=test")
+            .to_request();
         let resp = test::call_service(&app, req).await;
         assert!(resp.status().is_success());
         let bytes = to_bytes(resp.into_body()).await.unwrap();
@@ -459,7 +523,10 @@ mod tests {
     async fn evaluate_formula() {
         let conn = Connection::open_in_memory().unwrap();
         init_db(&conn);
-        let data = web::Data::new(AppState { db: Mutex::new(conn), sessions: Arc::new(Mutex::new(HashMap::new())) });
+        let data = web::Data::new(AppState {
+            db: Mutex::new(conn),
+            sessions: Arc::new(Mutex::new(HashMap::new())),
+        });
         let app = test::init_service(
             App::new()
                 .app_data(data.clone())
@@ -469,7 +536,10 @@ mod tests {
 
         let req = test::TestRequest::post()
             .uri("/evaluate")
-            .set_json(&EvalRequest { expr: "=SUM(1,2,3)".into() })
+            .set_json(&EvalRequest {
+                expr: "=SUM(1,2,3)".into(),
+                sheet: Some("test".into()),
+            })
             .to_request();
         let resp = test::call_service(&app, req).await;
         assert!(resp.status().is_success());
@@ -481,7 +551,10 @@ mod tests {
     async fn evaluate_average() {
         let conn = Connection::open_in_memory().unwrap();
         init_db(&conn);
-        let data = web::Data::new(AppState { db: Mutex::new(conn), sessions: Arc::new(Mutex::new(HashMap::new())) });
+        let data = web::Data::new(AppState {
+            db: Mutex::new(conn),
+            sessions: Arc::new(Mutex::new(HashMap::new())),
+        });
         let app = test::init_service(
             App::new()
                 .app_data(data.clone())
@@ -491,7 +564,10 @@ mod tests {
 
         let req = test::TestRequest::post()
             .uri("/evaluate")
-            .set_json(&EvalRequest { expr: "=AVERAGE(2,4,6)".into() })
+            .set_json(&EvalRequest {
+                expr: "=AVERAGE(2,4,6)".into(),
+                sheet: Some("test".into()),
+            })
             .to_request();
         let resp = test::call_service(&app, req).await;
         assert!(resp.status().is_success());
@@ -503,7 +579,10 @@ mod tests {
     async fn set_formula_cell() {
         let conn = Connection::open_in_memory().unwrap();
         init_db(&conn);
-        let data = web::Data::new(AppState { db: Mutex::new(conn), sessions: Arc::new(Mutex::new(HashMap::new())) });
+        let data = web::Data::new(AppState {
+            db: Mutex::new(conn),
+            sessions: Arc::new(Mutex::new(HashMap::new())),
+        });
         let app = test::init_service(
             App::new()
                 .app_data(data.clone())
@@ -512,7 +591,15 @@ mod tests {
         )
         .await;
 
-        let new_cell = Cell { row: 1, col: 1, value: "=SUM(2,3)".into(), font_weight: None, font_style: None, background_color: None };
+        let new_cell = Cell {
+            sheet: Some("test".into()),
+            row: 1,
+            col: 1,
+            value: "=SUM(2,3)".into(),
+            font_weight: None,
+            font_style: None,
+            background_color: None,
+        };
         let req = test::TestRequest::post()
             .uri("/cells")
             .set_json(&new_cell)
@@ -520,7 +607,9 @@ mod tests {
         let resp = test::call_service(&app, req).await;
         assert!(resp.status().is_success());
 
-        let req = test::TestRequest::get().uri("/cells").to_request();
+        let req = test::TestRequest::get()
+            .uri("/cells?sheet=test")
+            .to_request();
         let resp = test::call_service(&app, req).await;
         let bytes = to_bytes(resp.into_body()).await.unwrap();
         let cells: Vec<Cell> = serde_json::from_slice(&bytes).unwrap();
